% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{bye_weeks}
\alias{bye_weeks}
\title{Get bye weeks}
\usage{
bye_weeks(season = NULL, gameWeek = NULL)
}
\arguments{
\item{season}{The season that bye weeks are requested for}

\item{gameWeek}{The week that bye weeks are requested for. Leave blank to get
full season.}
}
\description{
Finds bye weeks based on the NFL schedule that is loaded into MFL.
}
