% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{free_agents}
\alias{free_agents}
\title{Free agents for active league}
\usage{
free_agents(position = NULL)
}
\arguments{
\item{position}{Position to return free agents for}
}
\description{
Lists free agent for th active league. Will take into account the league unit.
}
