% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\docType{data}
\name{league_app}
\alias{league_app}
\title{Basic league object that is generated at package startup}
\format{
An object of class \code{league} (inherits from \code{mfl_app}, \code{R6}) of length 65.
}
\usage{
league_app
}
\description{
Basic league object that is generated at package startup
}
\keyword{datasets}
