% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{league_schedule}
\alias{league_schedule}
\title{Schedule for active league}
\usage{
league_schedule(week = NULL, franchise = NULL)
}
\arguments{
\item{week}{The week to get the schedule for}

\item{franchise}{The franchise to get schedule for}
}
\description{
Lists schedule for the active league
}
