% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\docType{data}
\name{mfl_app}
\alias{mfl_app}
\title{Generic endpoint object}
\format{
An object of class \code{mfl_app} (inherits from \code{R6}) of length 15.
}
\usage{
mfl_app
}
\description{
The endpoint is generated at package startup and provides the basis for the
functions in the package.
}
\keyword{datasets}
