% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{mfl_login}
\alias{mfl_login}
\title{Login to MFL}
\usage{
mfl_login(user_name, password = NULL)
}
\arguments{
\item{user_name}{User name for user}

\item{password}{Password for user}
}
\description{
Logs a user into MFL and enables user specific data to be retreieved.
}
