% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{my_leagues}
\alias{my_leagues}
\title{List users leagues}
\usage{
my_leagues(franchise_names = FALSE)
}
\description{
Returns a list of leagues that belong to the user that is logged in via \link{mfl_login}.
}
