% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{nfl_schedule}
\alias{nfl_schedule}
\title{Get NFL Schedule}
\usage{
nfl_schedule(season = NULL, gameWeek = NULL)
}
\arguments{
\item{season}{Season that schedule should be retreived for.}

\item{gameWeek}{The week that the NFL schedule shoud be retreived for. Use
\code{"ALL"} to retrieve full season}
}
\value{
List of lists with matchups
}
\description{
Retreives NFL Schedule from MFL
}
