% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{player_profile}
\alias{player_profile}
\title{Player Profile}
\usage{
player_profile(player_id)
}
\arguments{
\item{player_id}{A single player ID or vector of multiple player IDs to get information on}
}
\description{
Returns a list with summary of information regarding a player, including DOB,
ADP ranking, height/weight.
}
