% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{player_ranks}
\alias{player_ranks}
\title{Player Ranks}
\usage{
player_ranks(position = NULL)
}
\arguments{
\item{position}{Filters the rankings to this position only (optional)}
}
\description{
Overall player rankings from the experts at FantasySharks.com. These rankings
can be used instead of Average Draft Position (ADP) rankings for guidance during
your draft, or when generating your own draft list.
}
