% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{player_scores}
\alias{player_scores}
\title{Player scores for active league}
\usage{
player_scores(week = NULL, season = NULL, player_id = NULL,
  position = NULL, status = NULL, rules = NULL, count = NULL)
}
\description{
Player scores for active league
}
