% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{players}
\alias{players}
\title{Player list}
\usage{
players(details = FALSE, since = NULL, player_id = NULL)
}
\arguments{
\item{details}{\code{TRUE/FALSE} indicating whether full detailed information
should be retreived}

\item{since}{Unix Timestamp to retrieve only changes to the player database
since that time}

\item{player_id}{Player IDs to retrieve information for}
}
\value{
Tibble with player information
}
\description{
Get list of players from MFL
}
