% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{points_allowed}
\alias{points_allowed}
\title{Fantasy points allowed}
\usage{
points_allowed()
}
\description{
Fantasy points allowed by each NFL team based on current league scoring
settings, broken out by position.
}
