% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{projected_scores}
\alias{projected_scores}
\title{Projected Scores for active league}
\usage{
projected_scores(
  week = NULL,
  player_id = NULL,
  position = NULL,
  status = NULL,
  count = NULL
)
}
\arguments{
\item{week}{The week to get scores from}

\item{player_id}{Player Id(s) for players to get projected scores for}

\item{position}{Position to get projected scores for}

\item{status}{Set to \code{"freeagent"} to only return projected scores for
free agents}

\item{count}{Set to limit number of players returned}
}
\description{
Get projected scores based on active leagues scoring rules.
}
