% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_rules.R
\name{scoring_rule}
\alias{scoring_rule}
\title{Base R6 Class for scoring rules}
\description{
Stores base information for a single scoring rule
}
\details{
Include method for calculating scoring.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{event}}{Stores the scoring event for the rule}

\item{\code{range}}{Stores the range for the event that is valid for the rule}

\item{\code{points}}{Stores how many points is given for the event. If first character
is \code{*} then it is a multiplier rule. If the points is in the format
\code{x/y} then x points is given for every y. If points is just a number then
that is what is scored if the value is within the range.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-scoring_rule-new}{\code{scoring_rule$new()}}
\item \href{#method-scoring_rule-print}{\code{scoring_rule$print()}}
\item \href{#method-scoring_rule-value}{\code{scoring_rule$value()}}
\item \href{#method-scoring_rule-match}{\code{scoring_rule$match()}}
\item \href{#method-scoring_rule-clone}{\code{scoring_rule$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scoring_rule-new"></a>}}
\if{latex}{\out{\hypertarget{method-scoring_rule-new}{}}}
\subsection{Method \code{new()}}{
Create a new rule
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{scoring_rule$new(event = NULL, range = NULL, points = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{event}}{The scoring event to create the rule for.}

\item{\code{range}}{The event range to create the scoring rule for.}

\item{\code{points}}{The points setting for the scoring rule.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Creates a new rule based on the event, range and points specified.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scoring_rule-print"></a>}}
\if{latex}{\out{\hypertarget{method-scoring_rule-print}{}}}
\subsection{Method \code{print()}}{
Displays Rule Details
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{scoring_rule$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scoring_rule-value"></a>}}
\if{latex}{\out{\hypertarget{method-scoring_rule-value}{}}}
\subsection{Method \code{value()}}{
Calculates the value resulting from the value x happening
for the event.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{scoring_rule$value(x, season = FALSE, season_weeks = 16)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The value for the event that needs to be scored.}

\item{\code{season}}{Is the value x for an entire season?}

\item{\code{season_weeks}}{Number of weeks included in the season.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scoring_rule-match"></a>}}
\if{latex}{\out{\hypertarget{method-scoring_rule-match}{}}}
\subsection{Method \code{match()}}{
Determines if supplied event and range matches with the current rule.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{scoring_rule$match(event, range)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{event}}{Which event to look for}

\item{\code{range}}{Which range to look for}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Can be used to find a specific rule
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scoring_rule-clone"></a>}}
\if{latex}{\out{\hypertarget{method-scoring_rule-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{scoring_rule$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
