% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\docType{data}
\name{season_state}
\alias{season_state}
\title{Season Information}
\format{
An object of class \code{list} of length 9.
}
\usage{
season_state
}
\description{
Information on current state of season from Sleeper.
}
\keyword{datasets}
