% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.R
\name{set_league}
\alias{set_league}
\title{Identify league}
\usage{
set_league(league_id, season = NULL)
}
\arguments{
\item{league_id}{Id of the league to use}

\item{season}{season to use if different from current season}
}
\description{
Sets the league id to be used.
}
