% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\docType{data}
\name{mfl_endpoint}
\alias{mfl_endpoint}
\title{MFL Endpoint}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
mfl_endpoint
}
\description{
R6 Class representing the the endpoint for MFL. Used for low level operations
such as logging in and generating request urls.
}
\keyword{datasets}
