% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{mfl_endpoint}
\alias{mfl_endpoint}
\title{MFL Endpoint}
\description{
R6 Class representing the the endpoint for MFL. Used for low level operations
such as logging in and generating request urls.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{mfl_endpoint$new()}}
\item \href{#method-print}{\code{mfl_endpoint$print()}}
\item \href{#method-login}{\code{mfl_endpoint$login()}}
\item \href{#method-export_query}{\code{mfl_endpoint$export_query()}}
\item \href{#method-import_query}{\code{mfl_endpoint$import_query()}}
\item \href{#method-request}{\code{mfl_endpoint$request()}}
\item \href{#method-export_url}{\code{mfl_endpoint$export_url()}}
\item \href{#method-import_url}{\code{mfl_endpoint$import_url()}}
\item \href{#method-logged_in}{\code{mfl_endpoint$logged_in()}}
\item \href{#method-clone}{\code{mfl_endpoint$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfl_endpoint$new(
  scheme = "https",
  host = "api.myfantasyleague.com",
  season = "2021"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfl_endpoint$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-login"></a>}}
\if{latex}{\out{\hypertarget{method-login}{}}}
\subsection{Method \code{login()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfl_endpoint$login(user_name, password = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-export_query"></a>}}
\if{latex}{\out{\hypertarget{method-export_query}{}}}
\subsection{Method \code{export_query()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfl_endpoint$export_query(qry_type, qry_params = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-import_query"></a>}}
\if{latex}{\out{\hypertarget{method-import_query}{}}}
\subsection{Method \code{import_query()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfl_endpoint$import_query(qry_type, qry_params = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-request"></a>}}
\if{latex}{\out{\hypertarget{method-request}{}}}
\subsection{Method \code{request()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfl_endpoint$request(request_url)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-export_url"></a>}}
\if{latex}{\out{\hypertarget{method-export_url}{}}}
\subsection{Method \code{export_url()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfl_endpoint$export_url()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-import_url"></a>}}
\if{latex}{\out{\hypertarget{method-import_url}{}}}
\subsection{Method \code{import_url()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfl_endpoint$import_url()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-logged_in"></a>}}
\if{latex}{\out{\hypertarget{method-logged_in}{}}}
\subsection{Method \code{logged_in()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfl_endpoint$logged_in()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mfl_endpoint$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
